backend/requirements.txt
fastapi
uvicorn[standard]                  # FastAPI ASGI server
sentence-transformers             # Embedding model
pymupdf                           # PDF extraction (PyMuPDF)
python-multipart                  # Form/file uploads
pymongo                           # MongoDB client
faiss-cpu                         # Vector similarity search
ollama                            # Local model interface
requests                          # For downloading files
beautifulsoup4                    # For parsing HTML emails
python-docx                       # For extracting DOCX files
